Clean Up and Move to New Workflow
Instead of primitively having a dozen files all over the place and a
mess of attempts and code, I'm going to try to make use of the tools
available, namely, maintain a master and branch off of that to add work,
then merge back. Hopefully always having a functional and complete
master.