Lots and lots and lots.
* SelectK on Session Data
Extracted top features from the session data using fclassif (150 features).

* RF with big gain
Used all features in RF - best yet performance locally and on LB.

* Feature Importance Examination
Looked at which features are important. It's a bit of a mess.
Not clear why "tfahour" should be at the top. And why are all the age_gender_brkts mattering so much?

* Validation Curve
Performance doesn't seem to change with parameters.
* Learning Curve
Performance doesn't change with number of training examples.